doc/ 2002-01-22  Marcus Brinkmann  <marcus@g10code.de>
	* gpgme.texi (Creating data buffers): Change
	gpgme_data_new_from_filepart's LENGTH type from off_t to size_t.

gpgme/
2002-01-22  Marcus Brinkmann  <marcus@g10code.de>

	* data.c (gpgme_data_new_from_filepart): Change type of LENGTH
	from off_t to size_t.
	* gpgme.h: Likewise.

+++ gpgme/data.c
@@ -283,7 +283,7 @@ gpgme_data_new_from_file (GpgmeData *r_dh, const char *fname, int copy)
  **/
 GpgmeError
 gpgme_data_new_from_filepart (GpgmeData *r_dh, const char *fname, FILE *fp,
-			      off_t offset, off_t length)
+			      off_t offset, size_t length)
 {
   GpgmeData dh;
   GpgmeError err;
